Display works after refactoring

+++ src/display/font.hpp
@@ -6,8 +6,8 @@ namespace display
 
 struct Font
 {
-    const uint8_t height;
-    const uint8_t width;
+    const uint16_t height;
+    const uint16_t width;
     const uint8_t* array;
 };
